Replace null characters with whitespaces in command lines
In linux / unix environments, command lines are parsed by the shell and used to launch the process with its argument list broken down into null terminated strings.

When reading the command line from procfs, the null character is also used to separate each argv[i].

Since this is inconvenient for searching processes, I decided to replace the null character with a whitespace character when readling the command line.