add getPointer() to Dereferencable
Some internal structures like zend_object or zend_string uses C struct hack heavily. And the implementation of the memory profiler often needs to know the remote address and it's size of structures. So it's convenient if we can always get the original remote address and size used to dereference the objects.